debugging fixes to db_registry poolsizes (accurate to retail) and debugging fixup to jobqueue area.
Jobqueue was kinda aids

+++ src/universal/q_shared.h
@@ -49,7 +49,7 @@
 int __cdecl ShortSwap(__int16 l);
 int __cdecl LongSwap(int l);
 
-static ID_INLINE short BigShort(short l) { return ShortSwap(l); }
+static ID_INLINE short BigShort(short l) { return (short)ShortSwap(l); }
 #define LittleShort
 static ID_INLINE int BigLong(int l) { return LongSwap(l); }
 #define LittleLong
